📦 NEW: Changed app name
