+++ wallpaper-generator/next.config.js
