+++ wallpaper-generator/pages/_app.tsx
