+++ wallpaper-generator/public/favicon.ico
