+++ wallpaper-generator/styles/globals.css
