+++ wallpaper-generator/tailwind.config.js
