+++ wallpaper-generator/tsconfig.json
